Adjust with subm-operator api rename
Depends on: https://github.com/submariner-io/submariner-operator/pull/1651
